+++ _includes/source_code/code/24-PDP/souvlakia/souvlakia_nlogn_implicit_st_cptr.cc
@@ -5,6 +5,7 @@
 #include <cstdio>
 #include <algorithm>
 #include <climits>
+#include <cstdint>
 #include <cassert>
 #include <vector>
 #include <list>
@@ -15,18 +16,18 @@
 
 using namespace std;
 
-const long 
-	MAXN = long(1e5),	//max shops
-	MAXQ = long(5e4),	//max queries
-	INF  = LONG_MAX;	//INF for dijkstra
-long	N,M,Q,C[3];
+const int32_t
+	MAXN = int32_t(1e5),	//max shops
+	MAXQ = int32_t(5e4),	//max queries
+	INF  = INT32_MAX;	//INF for dijkstra
+int32_t	N,M,Q,C[3];
 
 struct shop {
 	#define X	d[0]
 	#define Y 	d[1]
 	#define Z	d[2]
-	long d[3];//The 3 distances (X,Y,Z) from {A,B,C}
-	list<long> queryid;//query ids that has to answer shop[i]
+	int32_t d[3];//The 3 distances (X,Y,Z) from {A,B,C}
+	list<int32_t> queryid;//query ids that has to answer shop[i]
 			//use list in case there are more than one queries per shop
 	shop(){ X=Y=Z=INF; }//help dijsktra initial values
 	bool operator <(const shop& b) const { 
@@ -34,10 +35,10 @@ struct shop {
 	}
 } S[MAXN+1];
 
-void dijsktra(long src,long dupd,vector<vector<pair<long,long>>>& edge){
+void dijsktra(int32_t src,int dupd,vector<vector<pair<int32_t,int32_t>>>& edge){
 	//src is one of {A,B,C}
 	//dupd in [0,3). Update Shop[*].d[dupd] distance from src
-	set<pair<long,long>> DS;//<distance,shopid>
+	set<pair<int32_t,int32_t>> DS;//<distance,shopid>
 	
 	DS.insert({0,src});
 	S[src].d[dupd] = 0;
@@ -45,9 +46,9 @@ void dijsktra(long src,long dupd,vector<vector<pair<long,long>>>& edge){
 	while(!DS.empty()){//explore
 		auto x = *(DS.begin());
 		DS.erase(DS.begin());
-		long daddy = x.second;
+		int32_t daddy = x.second;
 		for(auto y : edge[daddy]){//expand
-			long child = y.first, dist = y.second;
+			int32_t child = y.first, dist = y.second;
 			if(S[daddy].d[dupd]!=INF && S[child].d[dupd] > S[daddy].d[dupd] + dist){
 				if(S[child].d[dupd]!=INF)
 					DS.erase(DS.find({S[child].d[dupd],child}));
@@ -58,8 +59,8 @@ void dijsktra(long src,long dupd,vector<vector<pair<long,long>>>& edge){
 	}
 }
 
-inline long read_fast(){//fast IO is required for this problem due to huge io
-	long x = 0;
+inline int32_t read_fast(){//fast IO is required for this problem due to huge io
+	int32_t x = 0;
 	char c;
 	while((c=getchar_unlocked())<'0');//skip blanks
 	do
@@ -75,18 +76,18 @@ inline void write_fast(bool f){
 }
 
 //RMQ with dynamic memory allocation for implicit segment tree
-long YMAX = long(2e4)*MAXN;	//max value of Y
+int32_t YMAX = int32_t(2e4)*MAXN;	//max value of Y
 
 struct node {//implicit segment tree (lazy build)
-	long left,right; //Y value limits that this node handles
-	long minZ; //min Z value of this node (and beyond)
+	int32_t left,right; //Y value limits that this node handles
+	int32_t minZ; //min Z value of this node (and beyond)
 	node *leftptr, *rightptr; //pointers to the children 
 		
-	node(long le,long ri, long z){
+	node(int32_t le,int32_t ri, int32_t z){
 		left = le, right = ri, minZ = z;
 		leftptr = rightptr = 0;
 	}
-	long query(long y){
+	int32_t query(int32_t y){
 		if(y < left)//not involved
 			return INF;
 		if(right <= y)//completely involved
@@ -97,7 +98,7 @@ struct node {//implicit segment tree (lazy build)
 		return min(leftptr->query(y),rightptr->query(y));
 		
 	}
-	void update(long y,long z){
+	void update(int32_t y,int32_t z){
 		expand();//preserve prev value for children (like lazy push ST)
 		minZ = min(minZ,z);
 		if(right == y)
@@ -111,7 +112,7 @@ struct node {//implicit segment tree (lazy build)
 	}
 	void expand(){
 		if(!leftptr && left<right){
-			long mid = (left + right)/2;
+			int32_t mid = (left + right)/2;
 			leftptr = new node(left,mid,minZ);
 			rightptr= new node(mid+1,right,minZ);
 		}
@@ -119,8 +120,8 @@ struct node {//implicit segment tree (lazy build)
 } root(0,YMAX,INF);
 
 void sinit(){}
-inline long squery(long y){ return root.query(y); }
-inline void supdate(long y,long z){ root.update(y,z); }
+inline int32_t squery(int32_t y){ return root.query(y); }
+inline void supdate(int32_t y,int32_t z){ root.update(y,z); }
 
 int main(){
 #ifdef CONTEST
@@ -130,25 +131,25 @@ int main(){
 	N = read_fast(), M = read_fast();
 
 	{//hint: edge vector below is temporary 
-		//max mem usage: 400.000 * 2 * 3 * sizeof(long) = 2.400.000 * sizeof(long) = 19.2Mb
+		//max mem usage: 400.000 * 2 * 3 * sizeof(int32_t) = 2.400.000 * sizeof(int32_t) = 19.2Mb
 		//and we definitely benefit our program by releasing this mem when no more required
 		//if we use a dynamic allocated scheme for implicit segment tree or use map template
-		vector<vector<pair<long,long>>> edge(N+1);//<edge_to, distance>
-		for(long a,b,d,i=0;i<M;++i){
+		vector<vector<pair<int32_t,int32_t>>> edge(N+1);//<edge_to, distance>
+		for(int32_t a,b,d,i=0;i<M;++i){
 			a = read_fast(), b = read_fast(), d = read_fast();
 			edge[a].push_back({b,d});
 			edge[b].push_back({a,d});
 		}
-		for(long i=0;i<3;i++)
+		for(int i=0;i<3;i++)
 			C[i]=read_fast();
 		Q = read_fast();
 
-		for(long i=0;i<3;++i)
+		for(int i=0;i<3;++i)
 			dijsktra(C[i],i,edge);
 	}//memory of edge vector is released
 
 	//read queries and store reference so we can answer offline
-	for(long i=0;i<Q;++i)
+	for(int32_t i=0;i<Q;++i)
 		S[read_fast()].queryid.push_back(i);//shop[q] has to answer i-th query
 
 	sort(S+1,S+N+1);
@@ -156,7 +157,7 @@ int main(){
 	vector<bool> ans(MAXQ+1,false);		//offline answers
 
 	sinit();
-	for(long right = 1, left = 1; left <=N;){
+	for(int32_t right = 1, left = 1; left <=N;){
 		while(right<=N && S[right].X == S[left].X){
 			//for all shops with equal X postpone updates
 			const shop& s = S[right];
@@ -172,7 +173,7 @@ int main(){
 		}
 	}
 	
-	for(long i=0;i<Q;++i)
+	for(int32_t i=0;i<Q;++i)
 		write_fast(ans[i]);
 
 	return 0;
